requirements.txt
absl-py==1.4.0
astunparse==1.6.3
cachetools==5.3.0
cffi==1.15.1
charset-normalizer==3.0.1
cloudpickle==2.2.1
contourpy==1.0.7
cycler==0.11.0
Cython==0.29.33
dm-control==0.0.425341097
dm-env==1.6
dm-tree==0.1.8
fasteners==0.18
flatbuffers==23.1.21
fonttools==4.38.0
future==0.18.3
gast==0.5.3
glfw==2.5.6
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
gtimer==1.0.0b5
gym==0.22.0
gym-notices==0.0.8
h5py==3.8.0
idna==3.4
imageio==2.25.0
importlib-metadata==6.0.0
joblib==1.2.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
kornia==0.6.9
labmaze==1.0.6
libclang==15.0.6.1
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.6.3
mujoco-py==2.1.2.14
numpy==1.23.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==23.0
Pillow==9.4.0
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyOpenGL==3.1.6
pyparsing==2.4.7
python-dateutil==2.8.2
python-utils==3.5.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.1
scipy==1.10.0
six==1.16.0
sklearn==0.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.2.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
torch==1.13.1
torchvision==0.14.1
tqdm==4.47.0
typing_extensions==4.4.0
urllib3==1.26.14
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.12.1
